pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "qibo-cloud-backends"
version = "0.0.5"
description = "Qibo backends for client interface."
authors = ["The Qibo team"]
license = "Apache License 2.0"
readme = "README.md"
homepage = "https://qibo.science/"
repository = "https://github.com/qiboteam/qibo-cloud-backends/"
documentation = "https://qibo.science/qibo-cloud-backends/stable"
keywords = []
classifiers = [
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Physics",
]
packages = [{ include = "qibo_cloud_backends", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
qibo = "^0.2.14"
qibo_client = "^0.0.15"
qiskit_ibm_runtime = { version = "^0.17", optional = true }
qiskit_ibm_provider = { version = "^0.11", optional = true }
amazon-braket-sdk = { version = "^1.83.0", optional = true }

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pylint = "^3.0.3"
qiskit_ibm_runtime = "^0.17"
qiskit_ibm_provider = "^0.11"
amazon-braket-sdk = "^1.83.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
recommonmark = "^0.7.1"
sphinx-copybutton = "^0.5.2"
nbsphinx = "^0.9.3"
furo = "^2023.9.10"
qiskit_ibm_runtime = "^0.17"
qiskit_ibm_provider = "^0.11"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ipython = "^7"
pdbpp = "^0.10.3"

[tool.poetry.extras]
aws = ["amazon-braket-sdk"]

[tool.poe.tasks]
test = "pytest"
lint = "pylint --errors-only src"
lint-warnings = "pylint --exit-zero src"
docs = "make -C doc html"
docs-clean = "make -C doc clean"
test-docs = "make -C doc doctest"

[tool.pylint.reports]
output-format = "colorized"

[tool.pytest.ini_options]
testpaths = ['tests/']
filterwarnings = ['ignore::RuntimeWarning']
addopts = [
  '--cov=src/qibo_cloud_backends',
  '--cov-report=xml',
  '--cov-report=html',
]
